content-creation-backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# 认证安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP客户端
httpx==0.25.2

# 异步任务
celery==5.3.4
redis==5.0.1

# 文件存储
minio==7.2.3
oss2>=2.18.0

# 工具库
python-dotenv==1.0.0
tenacity==8.2.3

# 日志
structlog==23.2.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 开发工具
black==23.11.0
isort==5.12.0
mypy==1.7.1
